Enable uni-directional parameter bridge
This commit adds the feature that the parameter bridge for selective
topic passing between ROS1 and ROS2 can be used also in an
uni-directional manner.

Old behavior: Always bi-directional, thus messages could appear multiple
times, as e.g. on the ROS2 side the ROS2 publisher and the bridge can
publish on the same topic.